+++ samples/520 - 2D Graph Traversal/main.cpp
@@ -157,7 +157,7 @@ public:
 	}
 	bool OnProcess()
 	{
-		unsigned long t = GetRunTimeMS();
+		unsigned long long t = GetRunTimeMS();
 		if((t - m_lastProcess) > 15)
 		{
 			float dt = (float)(t - m_lastProcess) / 1000.f;
