+++ samples/551 - 2D Server/main.cpp
@@ -166,7 +166,7 @@ public:
 	}
 	bool OnProcess()
 	{
-		unsigned long t = GetRunTimeMS();
+		unsigned long long t = GetRunTimeMS();
 		if((t - m_lastProcess) > 15)
 		{
 			float dt = (float)(t - m_lastProcess) / 1000.f;
@@ -272,13 +272,13 @@ void *clientFunc(void *arg)
 	string packetPrefix = "P:";
 	string readBuffer;
 
-	unsigned long lastPositionUpdate = app->GetRunTimeMS();
+	unsigned long long lastPositionUpdate = app->GetRunTimeMS();
 
 	unsigned short packetSize = sizeof(Packet);
 	int hasRead = 0;
 	while(true)
 	{
-		unsigned long t = app->GetRunTimeMS();
+		unsigned long long t = app->GetRunTimeMS();
 
 		hasRead = client->HasRead();
 		if(hasRead > 0)
